moved source into src and added images with readme
