+++ 10.0/AI/ChatClient/src/App.xaml
