+++ 10.0/AI/ChatClient/src/App.xaml.cs
