+++ 10.0/AI/ChatClient/src/AppShell.xaml
