+++ 10.0/AI/ChatClient/src/GlobalXmlns.cs
