+++ 10.0/AI/ChatClient/src/MauiProgram.cs
@@ -1,0 +1,36 @@
+﻿using Microsoft.Extensions.Logging;
+using SimpleChatClient.ViewModels;
+using SimpleChatClient.Services;
+using Fonts;
+
+namespace SimpleChatClient;
+
+public static class MauiProgram
+{
+	public static MauiApp CreateMauiApp()
+	{
+		var builder = MauiApp.CreateBuilder();
+		builder
+			.UseMauiApp<App>()
+			.ConfigureFonts(fonts =>
+			{
+				fonts.AddFont("OpenSans-Regular.ttf", "OpenSansRegular");
+				fonts.AddFont("OpenSans-Semibold.ttf", "OpenSansSemibold");
+				fonts.AddFont("FluentSystemIcons-Regular.ttf", FluentUI.FontFamily);
+			});
+
+		// Register ViewModels
+		builder.Services.AddSingleton<ChatViewModel>();
+
+		// Register AI chat client (Foundry/OpenAI) from environment variables (desktop only)
+#if WINDOWS || MACCATALYST
+		builder.AddFoundryChatClient();
+#endif
+
+#if DEBUG
+		builder.Logging.AddDebug();
+#endif
+
+		return builder.Build();
+	}
+}
