+++ 10.0/AI/ChatClient/src/Models/ChatMessage.cs
