+++ 10.0/AI/ChatClient/src/Platforms/Android/AndroidManifest.xml
