+++ 10.0/AI/ChatClient/src/Platforms/Android/MainActivity.cs
