+++ 10.0/AI/ChatClient/src/Platforms/Android/Resources/values/colors.xml
