+++ 10.0/AI/ChatClient/src/Platforms/MacCatalyst/AppDelegate.cs
