+++ 10.0/AI/ChatClient/src/Platforms/MacCatalyst/Entitlements.plist
