+++ 10.0/AI/ChatClient/src/Platforms/MacCatalyst/Info.plist
