+++ 10.0/AI/ChatClient/src/Platforms/MacCatalyst/Program.cs
