+++ 10.0/AI/ChatClient/src/Platforms/Windows/App.xaml
