+++ 10.0/AI/ChatClient/src/Platforms/Windows/Package.appxmanifest
