+++ 10.0/AI/ChatClient/src/Platforms/Windows/app.manifest
