+++ 10.0/AI/ChatClient/src/Platforms/iOS/AppDelegate.cs
