+++ 10.0/AI/ChatClient/src/Platforms/iOS/Info.plist
