+++ 10.0/AI/ChatClient/src/Platforms/iOS/Resources/PrivacyInfo.xcprivacy
