+++ 10.0/AI/ChatClient/src/Properties/launchSettings.json
