+++ 10.0/AI/ChatClient/src/Resources/Fonts/FluentSystemIcons-Regular.ttf
