+++ 10.0/AI/ChatClient/src/Resources/Fonts/FluentUI.cs
