+++ 10.0/AI/ChatClient/src/Resources/Fonts/OpenSans-Regular.ttf
