+++ 10.0/AI/ChatClient/src/Resources/Images/dotnet_bot.png
