+++ 10.0/AI/ChatClient/src/Resources/Raw/AboutAssets.txt
