+++ 10.0/AI/ChatClient/src/Resources/Styles/Colors.xaml
