+++ 10.0/AI/ChatClient/src/Resources/Styles/Styles.xaml
